backend/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-elevenlabs
livekit-plugins-deepgram
livekit-plugins-silero
python-dotenv
pyautogui
speedtest-cli
psutil
pycaw
screen-brightness-control
winshell
requests
pywhatkit
pyjokes
yfinance
google-api-python-client
google-auth-httplib2
google-auth-oauthlib